--- src/components/App/App.js
@@ -1,12 +1,0 @@
-import { StatusBar } from 'expo-status-bar';
-import SafeArea from '../SafeArea';
-import Navigation from '../Navigation';
-
-export default function App() {
-  return (
-    <SafeArea>
-        <Navigation />
-        <StatusBar style="auto" />
-    </SafeArea>
-  );
-}
