--- src/components/App/index.js
@@ -1,2 +1,0 @@
-export * from './App';
-export { default } from './App';
