+++ src/components/AuthProvider/index.js
@@ -1,0 +1,2 @@
+export * from './AuthProvider';
+export { default } from './AuthProvider';
